Updating publish step.
This change reworks the publish step to exclude files before copying them to the release folder, rather than after. This was needed to exclude the .vs folder, which would not be able to be copied and would throw errors if the project was open in visual studio.
I also changed the exclusion list to include git files, the readme file, and the .vs folder. The .vs folder is only relevant to people using visual studio.
And I also update it to prepare the released version as a zip, instead of a folder.